Move functions around, use wrapper for topwords
