+++ backend/Sender.ts
@@ -1,7 +1,5 @@
 import { Response } from 'express-serve-static-core';
-import sw from 'stopword';
-import stopwordsEn from 'stopwords-en';
-import { DatabaseMessageUtils, IMessage } from './database/Messages';
+import { DatabaseMessageUtils } from './database/Messages';
 import { DatabaseUserUtils } from './database/Users';
 import { DatabaseDuccUtils } from './database/DuccScores';
 import { LogWrapper } from './utils/logging/LogWrapper';
@@ -53,32 +51,6 @@ class Sender {
     }
   }
 
-  static async getTopWords(): Promise<void> {
-    const messages: IMessage[] = await DatabaseMessageUtils.getLinesLastNDays(7);
-    const wordCounts: Map<string, number> = new Map<string, number>();
-
-    for (const message of messages) {
-      const messageText: string = message.message.toLowerCase();
-
-      const splitAndCleanedWords: string[] = messageText
-        .split(/\s+/)
-        .map((word: string) => word.replace(/[^a-zA-Z0-9 ]/g, '').trim());
-
-      // More stop word sources are necessary to filter out all
-      const words = sw.removeStopwords(splitAndCleanedWords, [...sw.en, ...stopwordsEn]);
-
-      for (const word of words) {
-        if (word && !word.match(/^[0-9]*$/)) {
-          if (!wordCounts.has(word)) wordCounts.set(word, 1);
-          else wordCounts.set(word, wordCounts.get(word)! + 1);
-        }
-      }
-    }
-
-    const sortedWordCounts: Map<string, number> = new Map([...wordCounts.entries()].sort((a, b) => b[1] - a[1]));
-    await DatabaseMessageUtils.insertTopWords(sortedWordCounts);
-  }
-
   static async sendTopWords(res: Response<any, number>): Promise<void> {
     if (res) {
       log.debug('Sending top words to', { ip: res.req?.ip });
@@ -98,11 +70,5 @@ class Sender {
     }
   }
 }
-
-// every hour we do the top words calculation
-setInterval(() => {
-  log.info('Running top words calculation and inserting into DB');
-  Sender.getTopWords();
-}, 1000 * 60 * 60);
 
 export { Sender };
